Loading maps seems to work now. Parallax seems to work but need to write a lot of code to test it.

+++ include/map/MapLoader.h
@@ -35,7 +35,7 @@ namespace gravity
 
         void ParseTile(TileLayer* layer, TiXmlNode* parent);
         
-        int StringToNumber(const string &Text);
+        float StringToNumber(const string &Text);
 
         Game* game;
     };
